% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pyhon.R
\docType{data}
\name{caminho_conda}
\alias{caminho_conda}
\title{Author: Daniel A.M. Villela, Rafael Santos de Oliveira and Marlon Mesquita Lopes Cabreira
functions to implpement path signature in R
and lasso regression with signatures}
\format{
An object of class \code{character} of length 1.
}
\usage{
caminho_conda
}
\description{
Author: Daniel A.M. Villela, Rafael Santos de Oliveira and Marlon Mesquita Lopes Cabreira
functions to implpement path signature in R
and lasso regression with signatures
}
\keyword{datasets}
