% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.pathsig.R
\name{lasso.pathsig}
\alias{lasso.pathsig}
\title{Ajuste Padrão de Modelo Lasso com Assinaturas de Caminho}
\usage{
lasso.pathsig(L, y, m, Ltype = "list", family = "gaussian")
}
\arguments{
\item{L}{Lista ou matriz contendo os caminhos originais (séries temporais).}

\item{y}{Vetor contendo a variável resposta (target).}

\item{m}{Número inteiro indicando o nível de truncamento da assinatura.}

\item{Ltype}{String indicando o tipo de entrada (\code{"list"} por padrão).}

\item{family}{String indicando a família do modelo GLM (ex: \code{"gaussian"}, \code{"multinomial"}).}
}
\value{
Uma lista contendo o melhor valor de lambda (\code{lambda}), os coeficientes (\code{coef}),
  o modelo final ajustado (\code{model}), o modelo de validação cruzada (\code{cvmodel}),
  e a matriz de assinaturas gerada (\code{sig}).
}
\description{
Esta função ajusta um Modelo Linear Generalizado com penalização Lasso
utilizando as assinaturas de caminho extraídas dos dados originais.
Possui suporte explícito e otimizado para a família 'multinomial' (classificação
com múltiplas categorias), aplicando agrupamento aos coeficientes.
}
