% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadlag.R
\name{leadlag}
\alias{leadlag}
\title{Transformação de Caminho Lead-Lag (Produto de Kronecker)}
\usage{
leadlag(X)
}
\arguments{
\item{X}{Uma matriz ou array representando o caminho original.}
}
\value{
Uma matriz onde as colunas originais são duplicadas e deslocadas,
  retornando a estrutura combinada de avanço e atraso.
}
\description{
Esta é uma implementação alternativa e mais geral da transformação "lead-lag",
que utiliza a operação matemática do Produto de Kronecker (\code{\%x\%}) para
criar os desdobramentos de avanço (lead) e atraso (lag). Ela prepara matrizes
e arrays multidimensionais para a extração de assinaturas de caminho.
}
