% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadlag2.R
\name{leadlag2}
\alias{leadlag2}
\title{Transformação de Caminho Lead-Lag}
\usage{
leadlag2(X)
}
\arguments{
\item{X}{Uma matriz ou data frame contendo o caminho original. Espera-se
que tenha pelo menos duas colunas (representando as duas dimensões a
serem transformadas).}
}
\value{
Um array (matriz) bidimensional contendo o caminho transformado
com a técnica lead-lag.
}
\description{
Esta função aplica a transformação "lead-lag" (avanço-atraso) a um caminho
bidimensional. Essa técnica é frequentemente usada no cálculo de assinaturas
de caminho para capturar a variação quadrática e entender dependências
temporais nos dados. Ela intercala e duplica os pontos criando um caminho em degraus.
}
