% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_python.R
\name{path_signature}
\alias{path_signature}
\title{Calculates the signature of a path using the Python module esig}
\usage{
path_signature(path, truncation = 2, env_name = "r-esig")
}
\arguments{
\item{truncation}{Maximum depth of the signature default = 2.}

\item{env_name}{Name of the Conda environment containing the esig module
usually the one created by setup_esig_env.}

\item{data}{Matrix or data frame where each row is a point on the path.
The columns represent the dimensions.}
}
\value{
A numeric vector with the signature coefficients up to the specified depth.
}
\description{
Calculates the signature of a path using the Python module esig
}
\examples{
\dontrun{
path <- matrix(c(0, 1, 1, 0, 0, 1), ncol = 2, byrow = TRUE)
path_signature(path, truncation = 2)
}
}
