% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathsig.esig.R
\name{pathsig.esig}
\alias{pathsig.esig}
\title{function to implement path signature}
\usage{
pathsig.esig(path, m)
}
\arguments{
\item{path}{define path array}

\item{m}{truncation level-integer}
}
\description{
function to implement path signature
}
\examples{
path <- array(c(1, 1, 3, 10, 20, 30), c(3, 2))
m<-2
pathsig.esig(path,m)

}
