% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathsig.logsig.R
\name{pathsig.logsig}
\alias{pathsig.logsig}
\title{function to implement log path signature}
\usage{
pathsig.logsig(path, m)
}
\arguments{
\item{path}{define path array}

\item{m}{truncation level-integer

log-signature}
}
\description{
function to implement log path signature
}
\examples{
path <- array(c(1, 1, 3, 10, 20, 30), c(3, 2))
m<-2
pathsig.logsig(path,m)

}
