% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathsig.one.R
\name{pathsig.one}
\alias{pathsig.one}
\title{function to implement path signature or log path signature}
\usage{
pathsig.one(path, m, logT = TRUE)
}
\arguments{
\item{path}{define path array}

\item{m}{truncation level-integer

logT = TRUE (log-signature) or logT = FALSE}
}
\description{
function to implement path signature or log path signature
}
\examples{
path <- array(c(1, 1, 3, 10, 20, 30), c(3, 2))
m<-2
pathsig.one(path,m, logT= TRUE)

}
