% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathsig.R
\name{pathsig}
\alias{pathsig}
\title{Aplicação em Lote de Assinaturas de Caminho}
\usage{
pathsig(L, Ltype, m)
}
\arguments{
\item{L}{A estrutura de dados contendo os caminhos. Pode ser uma lista (onde
cada elemento é um caminho) ou uma matriz.}

\item{Ltype}{Uma string de texto indicando o tipo de entrada. Se for \code{"list"},
a função itera sobre os elementos da lista. Caso contrário, itera sobre as colunas.}

\item{m}{Um número inteiro (\code{integer}) que define o nível de truncamento
(truncation level) para o cálculo da assinatura.}
}
\value{
Uma matriz transposta onde cada linha corresponde à assinatura
  calculada de um dos caminhos fornecidos na entrada.
}
\description{
Esta função atua como um facilitador (wrapper) para aplicar o cálculo de
assinaturas de caminho (\code{pathsig.one}) a múltiplos caminhos de uma só vez.
Ela pode processar caminhos que estão armazenados em uma lista ou nas colunas
de uma matriz/data frame.
}
