% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pyhon.R
\name{setup_esig_env}
\alias{setup_esig_env}
\title{Configura ambiente Conda com o módulo Python 'esig'}
\usage{
setup_esig_env(
  conda_path = NULL,
  env_name = "r-esig",
  python_version = "3.10",
  force_recria = FALSE
)
}
\arguments{
\item{conda_path}{Caminho para o executável do Conda. Se `NULL` (padrão),
o `reticulate` tentará localizar automaticamente. Recomenda-se informar
o caminho explicitamente para evitar problemas.}

\item{env_name}{Nome do ambiente Conda a ser usado/criado.}

\item{python_version}{Versão do Python para o novo ambiente (padrão `"3.10"`).}

\item{force_recria}{Se `TRUE`, recria o ambiente mesmo se ele já existir.}
}
\value{
Invisivelmente, o resultado de `reticulate::use_condaenv()`.
}
\description{
Cria (se necessário) um ambiente Conda chamado `"r-esig"`, instala as
dependências (`numpy`) via conda-forge e o pacote `esig` via pip.
Por fim, aponta o `reticulate` para usar esse ambiente.
}
\examples{
\dontrun{
setup_esig_env(conda_path = "C:/Users/User/Anaconda3/Scripts/conda.exe")
}
}
