% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timejoined.R
\name{timejoined}
\alias{timejoined}
\title{Transformação de Caminho Time-Joined}
\usage{
timejoined(X)
}
\arguments{
\item{X}{Um data frame ou matriz representando o caminho original.
Geralmente espera-se que tenha pelo menos duas colunas.}
}
\value{
Um data frame com duas colunas (\code{time} e \code{value}) contendo
o caminho transformado e estruturado para o pacote esig.
}
\description{
Esta função recebe um conjunto de dados sequenciais e aplica a transformação
matemática "time-joined" (tempo-junto). Ela intercala os valores criando uma
estrutura em degraus, o que é um passo de pré-processamento comum antes do
cálculo de assinaturas de caminho (path signatures).
}
\examples{
# Carregando o pacote
library(pathsignatures)

# Criando um caminho bidimensional simples (3 observações no tempo)
Xpath <- matrix(c(2, 3, 4, 5, 5, 6), ncol = 2)

# Exibindo o caminho original
Xpath

# Aplicando a transformação timejoined
tj <- timejoined(Xpath)

# Visualizando a nova matriz transformada em degraus
tj

}
