% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_python.R
\name{setup_esig_env}
\alias{setup_esig_env}
\title{Setup Conda environment}
\usage{
setup_esig_env(
  conda_path = NULL,
  env_name = "r-esig",
  python_version = "3.10",
  force_recria = FALSE
)
}
\arguments{
\item{conda_path}{Path to the Conda executable. If NULL default,
reticulate will try to locate it automatically. It is recommended to specify
the path explicitly to avoid problems.}

\item{env_name}{Name of the Conda environment to be used/created.}

\item{python_version}{Python version for the new environment default "3.10".}

\item{force_recreate}{If TRUE, recreates the environment even if it already exists.}
}
\value{
Invisibly, the result of reticulate::use_condaenv.
}
\description{
Configures a Conda environment with the Python module 'esig'
Creates, if necessary, a Conda environment called "r-esig"`, installs the
dependencies numpy via conda-forge and the esig package via pip.
Finally, points reticulate to use this environment.
}
\examples{
\dontrun{
setup_esig_env(conda_path = "C:/Users/User/Anaconda3/Scripts/conda.exe")
}
}
